function that produces output according to a format
