+++ _putchar.c
@@ -1,0 +1,8 @@
+#include "main.h"
+#include <stdio.h>
+
+int _putchar(char c)
+{
+    putchar(c);
+    return (0);
+}
