+++ main.h
@@ -1,0 +1,7 @@
+#ifndef _MAIN_H
+#define _MAIN_H
+
+int _putchar(char c);
+int _printf(const char *format, ...);
+
+#endif/*_printf*/
