--- practice.h
@@ -1,6 +1,0 @@
-#ifndef _PRACTICE_H
-#define _PRACTICE_H
-
-int _printf(const char *format, ...);
-
-#endif
